requirements.txt
streamlit==1.37.0
openai==1.37.1
anthropic==0.32.0
transformers==4.43.3
torch==2.4.0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.32.3
sqlalchemy==2.0.31